fix_out_table_failing_in_ADC

+++ nzpyida/analytics/transform/preparation.py
@@ -236,9 +236,9 @@ def train_test_split(in_df: IdaDataFrame, out_table_train: str=None, out_table_t
         else:
             id_column = in_df.indexer
 
-    if in_df._idadb.exists_table_or_view(out_table_train):
+    if out_table_train and in_df._idadb.exists_table_or_view(out_table_train):
             in_df._idadb.drop_table(out_table_train)
-    if in_df._idadb.exists_table_or_view(out_table_test):
+    if out_table_test and in_df._idadb.exists_table_or_view(out_table_test):
             in_df._idadb.drop_table(out_table_test)
 
     temp_view_name, need_delete = materialize_df(in_df)
